+++ src/components/accordion/query.test.ts
@@ -4,7 +4,7 @@
 
 import { ParagraphAccordion } from '@/types/drupal/paragraph'
 import { queries } from '@/data/queries'
-import mockData from '@/mocks/accordion.mock.json'
+import mockData from '@/components/accordion/mock.json'
 
 const AccordionMock: ParagraphAccordion = mockData
