+++ src/components/collapsiblePanel/mock.js
