+++ src/components/collapsiblePanel/query.test.ts
@@ -4,7 +4,7 @@
 
 import { ParagraphCollapsiblePanel } from '@/types/drupal/paragraph'
 import { queries } from '@/data/queries'
-import { mockResponse } from '@/mocks/collapsiblePanel.mock'
+import { mockResponse } from '@/components/collapsiblePanel/mock'
 import { params } from './query'
 const CollapsiblePanel: ParagraphCollapsiblePanel = mockResponse
