+++ src/components/collapsiblePanelItem/mock.js
