+++ src/components/collapsiblePanelItem/query.test.ts
@@ -4,7 +4,7 @@
 
 import { ParagraphCollapsiblePanelItem } from '@/types/drupal/paragraph'
 import { queries } from '@/data/queries'
-import { mockResponse } from '@/mocks/collapsiblePanelItem.mock'
+import { mockResponse } from '@/components/collapsiblePanelItem/mock'
 
 const CollapsiblePanelItem: ParagraphCollapsiblePanelItem = mockResponse
