+++ src/components/qaSection/query.test.ts
@@ -4,7 +4,7 @@
 
 import { ParagraphQaSection } from '@/types/drupal/paragraph'
 import { queries } from '@/data/queries'
-import mockData from '@/mocks/qaSection.mock.json'
+import mockData from '@/components/qaSection/mock.json'
 
 const QaSectionMock: ParagraphQaSection = mockData
