+++ src/components/reactWidget/mock.js
