+++ src/components/wysiwyg/query.test.ts
@@ -4,7 +4,7 @@
 
 import { queries } from '@/data/queries'
 import { ParagraphWysiwyg } from '@/types/drupal/paragraph'
-import mockData from '@/mocks/wysiwyg.mock.json'
+import mockData from '@/components/wysiwyg/mock.json'
 
 const wysiwygMock: ParagraphWysiwyg[] = mockData
